ENG7-4430: Harden emergency restore metadata storage (#634)
## Merge summary — Total Upkeep 1.17.3

Tracks ENG7-4430. Moves emergency restore metadata into the backup directory so it survives plugin updates, improves reliability around restore-info writes and migrations, and bundles platform/CI compatibility fixes.

### Emergency restore metadata
- Store restore metadata and the per-install CLI identifier in the backup directory (with durable locators under `wp-content`)
- Migrate legacy restore-info on upgrade; clean up obsolete plugin-tree runtime files after a successful migration
- Prefer the durable `wp-content` locator when plugin and content locators disagree
- Carry metadata across backup-directory changes; roll back locators/state cleanly when migration or setup fails
- Do not write restore-info under the plugin tree when backup-directory storage cannot be established
- Surface restore-info write problems in the In Progress UI, backup emails, cron Archiver task status, and logs (archive can still succeed)
- Align backup-dir writability checks (including a direct write probe / filesystem fallback) so restore-info writes succeed when the directory is usable
- Exclude restore locator / restore-info artifacts from site archives

### ZIP / archive browser
- Repair missing ZIP64 EOCD metadata on large archives so browse/extract works via ZipArchive, PclZip, and emergency CLI
- Safer repair verification (restore original trailer on failed/partial writes; avoid accepting empty/broken opens as repaired)
- Correct folder detection via ZipArchive external attributes

### WordPress / PHP compatibility
- Defer plugin bootstrap and translated strings until `init` (WP 6.7+ textdomain notices); keep Premium storage providers loading after deferred bootstrap
- Fix localization timing when running inside `init`
- Fix FTP disconnect on PHP 8 (`FTP\Connection`); fix Pure-FTPd `nlist` path matching for uploaded detection; quiet benign FTPS close warnings
- PHP minimum 7.4; BoldGrid Library registration fixed for Composer 2 `installed.json`

### CI / tooling
- Travis: Composer 2; test on PHP 7.4 and 8.5
- MySQL 8 grant parsing on Focal for VIEW backup/import tests
- Quiet PHP 8.5 deprecation noise in tests; PHPUnit 7-compatible `phpunit.xml`
- Preserve custom local cron interval labels when translating defaults

### Review follow-ups
Addressed Copilot and Bugbot feedback on metadata migration, locator durability, Archiver success handling, ZIP repair edge cases, and related regression fixes.

+++ tests/admin/migrate/test-class-boldgrid-backup-admin-migrate-util.php
@@ -76,7 +76,7 @@ class Test_Boldgrid_Backup_Admin_Migrate_Util extends WP_UnitTestCase {
 		// Create the large file to be split
 		$large_file_size = 1024 * 1024 * 5; // 5MB
 		$file_handle     = fopen( $this->test_file_path, 'wb' );
-		fwrite( $file_handle, str_repeat( 'Hello World!', $large_file_size / 100 ) );
+		fwrite( $file_handle, str_repeat( 'Hello World!', (int) ( $large_file_size / 100 ) ) );
 		fclose( $file_handle );
 
 		// Act
